+++ example/index.js
@@ -26,6 +26,10 @@ map.on('load', function() {
 
 geocoder.on('result', getResult);
 
+geocoder.on('results', function(e) {
+  console.log('results: ', e.results);
+});
+
 function getResult() {
   console.log('Fetched', geocoder.getResult());
   geocoder.off('result', getResult);